Move service from cmd/frontend/internal/own to cmd/frontend/backend/own*.go
