+++ cmd/frontend/backend/own.go
@@ -1,4 +1,4 @@
-package own
+package backend
 
 import (
 	"bytes"
@@ -12,21 +12,21 @@ import (
 	codeownerspb "github.com/sourcegraph/sourcegraph/internal/own/codeowners/proto"
 )
 
-// Service gives access to code ownership data.
+// OwnService gives access to code ownership data.
 // At this point only data from CODEOWNERS file is presented, if available.
-type Service interface {
+type OwnService interface {
 	// OwnersFile returns a CODEOWNERS file from a given repository at given commit ID.
 	// In the case the file can not be found, `nil` `*codeownerspb.File` and `nil` `error` is returned.
 	OwnersFile(context.Context, api.RepoName, api.CommitID) (*codeownerspb.File, error)
 }
 
-var _ Service = service{}
+var _ OwnService = ownService{}
 
-func NewService(git gitserver.Client) Service {
-	return service{git: git}
+func NewOwnService(git gitserver.Client) OwnService {
+	return ownService{git: git}
 }
 
-type service struct {
+type ownService struct {
 	git gitserver.Client
 }
 
@@ -40,7 +40,7 @@ var codeownersLocations = []string{
 	"docs/CODEOWNERS",
 }
 
-func (s service) OwnersFile(ctx context.Context, repoName api.RepoName, commitID api.CommitID) (*codeownerspb.File, error) {
+func (s ownService) OwnersFile(ctx context.Context, repoName api.RepoName, commitID api.CommitID) (*codeownerspb.File, error) {
 	var content []byte
 	var err error
 	for _, path := range codeownersLocations {
